test(api): split route test modules
